+++ packages/typespec-go/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.14.3 (unreleased)
 
+### Bugs Fixed
+
+* Fixed adapting `Access.internal` operations that return a polymorphic type.
+
 ### Other Changes
 
 * Updated to the latest tsp toolset.